adjust skipping to allow for eitehr log or archive to be present

+++ lib/jenkins/jobdir.rb
@@ -41,17 +41,14 @@ module Jenkins
 
       # Filter now locked directories
       content.reject! { |d| locked.include?(File.realpath(d)) }
-      # Filter directories that have no log anymore
-      content.reject! { |d| !File.exist?("#{d}/log") }
 
       content.sort_by! { |c| File.basename(c).to_i }
       content[0..-6].each do |d| # Always keep the last 6 builds.
         log = "#{d}/log"
         archive = "#{d}/archive"
         next unless age(log) > 14
-        FileUtils.rm(File.realpath(log))
-        archive_path = File.realpath(archive)
-        FileUtils.rm_rf(archive_path) if File.exist?(archive_path)
+        FileUtils.rm(File.realpath(log)) if File.exist?(log)
+        FileUtils.rm_r(File.realpath(archive)) if File.exist?(archive)
       end
     end
   end
